tests y faucet actualizados

+++ migrations/4_deploy_UwUtoken.js
@@ -1,5 +1,5 @@
 const UwUtoken = artifacts.require('UwUtoken');
 
 module.exports = function(deployer) {
-  deployer.deploy(UwUtoken, 10**6 * 10**18);
+  deployer.deploy(UwUtoken, BigInt(10**6 * 10**18));
 }
